handle multi-region bam files and zero coverage holes
