+++ libs/bam-api/CMakeLists.txt
@@ -1,9 +1,10 @@
 add_library(bam-api STATIC
-  src/bam_api.cpp
+  src/region_api.cpp
   src/read.cpp
   src/aos_paired_reads.cpp src/soa_paired_reads.cpp
   src/amplicon.cpp src/amplicon_set.cpp
-  src/bam_api_config_builder.cpp
+  src/bam_file_config_builder.cpp
+  src/bam_file_manager.cpp
 )
 
 target_include_directories(bam-api PUBLIC ${CMAKE_SOURCE_DIR}/libs/bam-api/include)
