+++ libs/bam-api/include/bam-api/amplicon.hpp
@@ -1,6 +1,7 @@
 #ifndef BAM_API_AMPLICONS_HPP
 #define BAM_API_AMPLICONS_HPP
 
+#include <string>
 #include "bam-api/read.hpp"
 
 namespace bam_api {
