+++ libs/qmcp-solver/include/qmcp-solver/quasi_mcp_cuda_max_flow_solver.hpp
@@ -8,7 +8,7 @@
 #include <memory>
 #include <vector>
 
-#include "bam-api/bam_api.hpp"
+#include "bam-api/region_api.hpp"
 #include "bam-api/paired_reads.hpp"
 #include "solver.hpp"
 
@@ -24,7 +24,7 @@ class QuasiMcpCudaMaxFlowSolver : public Solver {
 
     enum class EdgeDirection : uint8_t { Forward, Backward };
 
-    std::unique_ptr<Solution> solve(uint32_t required_cover, bam_api::BamApi& bam_api) override;
+    std::unique_ptr<Solution> solve(uint32_t required_cover, bam_api::RegionApi& reg_api) override;
     bool uses_quality_of_reads() override { return false; }
 
     void set_block_size(uint32_t block_size);
