+++ libs/qmcp-solver/include/qmcp-solver/solver.hpp
@@ -5,7 +5,7 @@
 #include <memory>
 #include <vector>
 
-#include "bam-api/bam_api.hpp"
+#include "bam-api/region_api.hpp"
 #include "bam-api/read.hpp"
 
 namespace qmcp {
@@ -15,7 +15,7 @@ typedef std::vector<bam_api::ReadIndex> Solution;
 class Solver {
    public:
     virtual ~Solver() = default;
-    virtual std::unique_ptr<Solution> solve(uint32_t max_coverage, bam_api::BamApi& bam_api) = 0;
+    virtual std::unique_ptr<Solution> solve(uint32_t max_coverage, bam_api::RegionApi& reg_api) = 0;
     virtual bool uses_quality_of_reads() = 0;
 };
 }  // namespace qmcp
