+++ libs/qmcp-solver/include/qmcp-solver/test_solver.hpp
@@ -1,5 +1,5 @@
 #include <memory>
-#include "bam-api/bam_api.hpp"
+#include "bam-api/region_api.hpp"
 #ifndef QMCP_TEST_SOLVER_HPP
 #define QMCP_TEST_SOLVER_HPP()
 
@@ -8,7 +8,7 @@
 namespace qmcp {
 class TestSolver : public Solver {
    public:
-    std::unique_ptr<Solution> solve(uint32_t max_coverage, bam_api::BamApi& bam_api) override;
+    std::unique_ptr<Solution> solve(uint32_t max_coverage, bam_api::RegionApi& reg_api) override;
     bool uses_quality_of_reads() override { return false; }
 };
 }  // namespace qmcp
