+++ libs/qmcp-solver/src/quasi_mcp_cpu_max_flow_solver.cpp
@@ -4,13 +4,13 @@
 #include <memory>
 #include <vector>
 
-#include "bam-api/bam_api.hpp"
+#include "bam-api/region_api.hpp"
 #include "logging/log.hpp"
 #include "qmcp-solver/solver.hpp"
 
 std::unique_ptr<qmcp::Solution> qmcp::QuasiMcpCpuMaxFlowSolver::solve(uint32_t max_coverage,
-                                                                      bam_api::BamApi& bam_api) {
-    input_sequence_ = bam_api.get_paired_reads_aos();
+                                                                      bam_api::RegionApi& reg_api) {
+    input_sequence_ = reg_api.get_paired_reads_aos();
 
     operations_research::SimpleMaxFlow max_flow;
