+++ libs/qmcp-solver/src/test_solver.cpp
@@ -2,14 +2,14 @@
 
 #include <memory>
 
-#include "bam-api/bam_api.hpp"
+#include "bam-api/region_api.hpp"
 #include "bam-api/read.hpp"
 #include "bam-api/soa_paired_reads.hpp"
 #include "qmcp-solver/solver.hpp"
 
 std::unique_ptr<qmcp::Solution> qmcp::TestSolver::solve(uint32_t max_coverage,
-                                                        bam_api::BamApi& bam_api) {
-    const bam_api::SOAPairedReads& paired_reads = bam_api.get_paired_reads_soa();
+                                                        bam_api::RegionApi& reg_api) {
+    const bam_api::SOAPairedReads& paired_reads = reg_api.get_paired_reads_soa();
 
     std::unique_ptr<Solution> solution = std::make_unique<Solution>();
     solution->reserve(paired_reads.get_reads_count());
