+++ src/tests/coverage_tester.cpp
@@ -8,6 +8,7 @@
 #include <random>
 
 #include "bam-api/read.hpp"
+#include "bam-api/region_api.hpp"
 #include "logging/log.hpp"
 #include "qmcp-solver/solver.hpp"
 #include "reads_gen.hpp"
@@ -119,10 +120,10 @@ bool CoverageTester::is_out_cover_valid(std::vector<uint32_t>& in_cover,
 CoverageTestResult CoverageTester::small_example_test(qmcp::Solver& solver) {
     const uint32_t m = 4;
     auto input = get_small_aos_example();
-    bam_api::BamApi bam_api(input);
-    auto input_cover = bam_api.find_input_cover();
-    auto output_indices = solver.solve(m, bam_api);
-    auto output_cover = bam_api.find_filtered_cover(*output_indices);
+    bam_api::RegionApi reg_api(input);
+    auto input_cover = reg_api.find_input_cover();
+    auto output_indices = solver.solve(m, reg_api);
+    auto output_cover = reg_api.find_filtered_cover(*output_indices);
     assert(is_out_cover_valid(input_cover, output_cover, m));
     return {input_cover, output_cover};
 }
@@ -137,10 +138,10 @@ CoverageTestResult CoverageTester::random_uniform_dist_test(qmcp::Solver& solver
     std::mt19937 mt(seed);
     auto input = reads_gen::rand_reads_uniform(mt, pairs_count, genome_length, read_length);
 
-    bam_api::BamApi bam_api(input);
-    auto input_cover = bam_api.find_input_cover();
-    auto output_indices = solver.solve(m, bam_api);
-    auto output_cover = bam_api.find_filtered_cover(*output_indices);
+    bam_api::RegionApi reg_api(input);
+    auto input_cover = reg_api.find_input_cover();
+    auto output_indices = solver.solve(m, reg_api);
+    auto output_cover = reg_api.find_filtered_cover(*output_indices);
     assert(is_out_cover_valid(input_cover, output_cover, m));
     return {input_cover, output_cover};
 }
@@ -156,10 +157,10 @@ CoverageTestResult CoverageTester::random_with_func_dist_test(
     std::mt19937 mt(seed);
     auto input = reads_gen::rand_reads(mt, pairs_count, genome_length, read_length, dist_func);
 
-    bam_api::BamApi bam_api(input);
-    auto input_cover = bam_api.find_input_cover();
-    auto output_indices = solver.solve(m, bam_api);
-    auto output_cover = bam_api.find_filtered_cover(*output_indices);
+    bam_api::RegionApi reg_api(input);
+    auto input_cover = reg_api.find_input_cover();
+    auto output_indices = solver.solve(m, reg_api);
+    auto output_cover = reg_api.find_filtered_cover(*output_indices);
     assert(is_out_cover_valid(input_cover, output_cover, m));
     return {input_cover, output_cover};
 }
